move Archipelago from domain to logic
